requirements.txt
requests>=2.31.0
python-telegram-bot>=20.0
pytz>=2023.3
schedule>=1.2.0
